[jk] Refactor OverwriteVariables table into separate component
